+++ releasenotes.md
@@ -1,5 +1,7 @@
 Whats New in v10.3
 --------------------------
+- Added JSON import for Making Formulas
+- Added JSON export for all and specific formulas in making
 - Layout update when adding suppliers for ingredients
 - Added JSON file import for ingredient and formula categories
 - Further improvements in SDS generation